backend/pyproject.toml
[project]
name = "lexarena-backend"
version = "0.1.0"
description = "FastAPI backend for LexArena legal simulation platform"
requires-python = ">=3.11"
dependencies = [
    "fastapi[all]>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.0",
    "supabase>=2.6.0",
    "crewai>=0.1.0",           # ADD THIS
    "langchain-openai>=0.0.5"  # ADD THIS
]

[project.optional-dependencies]
dev = ["ruff>=0.5.0", "mypy>=1.10.0"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"


